requirements.txt
asgiref==3.6.0
Django==3.2
django-cors-headers==3.13.0
django-crispy-forms==2.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
PyJWT==1.7.1
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
django-rest-auth==0.9.5